+++ docs/changelog.md
@@ -6,7 +6,7 @@ This page tracks releases of `earthmover`, with a summary of what was changed, f
 ### [v0.4.6](https://github.com/edanalytics/earthmover/releases/tag/v0.4.6)
 (Released 2025-08-14)
 
-* bugfix: `rename_columns` operation to an existing column name could result in two columns of the same name; now this results in an error
+* bugfix: []`rename_columns` operation to an existing column name could result in two columns of the same name; now this results in an error](https://github.com/edanalytics/earthmover/pull/166)
 * feature: [Add support for non-UTF8 file encodings for fixed-width inputs](https://github.com/edanalytics/earthmover/pull/165)