+++ docs/changelog.md
@@ -1,10 +1,17 @@
 This page tracks releases of `earthmover`, with a summary of what was changed, fixed, added in each new version.
 
-## Unreleased changes
-
-* feature: Add `melt` and `pivot` as dataframe operations
 
 ## 2025 releases
+
+### [v0.4.7](https://github.com/edanalytics/earthmover/releases/tag/v0.4.7)
+(Released 2025-11-13)
+
+* bugfix: [column jumbling when using `optional_columns`](https://github.com/edanalytics/earthmover/pull/171)
+* bugfix: [`earthmover init` failure when installed via pip](https://github.com/edanalytics/earthmover/pull/175)
+* bugfix: [use `removeprefix()` not `lstrip()` when processing `--set` flags](https://github.com/edanalytics/earthmover/pull/174)
+* bugfix: [prevent existing `value` column from being clobbered, if it exists](https://github.com/edanalytics/earthmover/pull/170)
+* feature: [add `melt` and `pivot` dataframe operations](https://github.com/edanalytics/earthmover/pull/168)
+
 
 ### [v0.4.6](https://github.com/edanalytics/earthmover/releases/tag/v0.4.6)
 (Released 2025-08-14)